add rounding to values fix NPE when date is not explicitly selected from picker in add expense view fix month+1

+++ app/src/main/java/com/matpaw/myexspencer/model/Limit.java
@@ -2,21 +2,22 @@ package com.matpaw.myexspencer.model;
 
 import android.support.annotation.NonNull;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
 public class Limit implements Comparable<Limit> {
     private UUID id;
     private Date date;
-    private float value;
+    private BigDecimal value;
 
-    public Limit(UUID id, Date date, float value) {
+    public Limit(UUID id, Date date, BigDecimal value) {
         this.id = id;
         this.date = date;
         this.value = value;
     }
 
-    public float getValue() {
+    public BigDecimal getValue() {
         return value;
     }
